+++ core/src/charts/iASPLOMData.h
@@ -52,8 +52,8 @@ public:
 	void setInverted(size_t paramIndex, bool isInverted);//!< Set whether the axis of a parameter should be inverted in the scatter plots.
 	void setParameterNames(std::vector<QString> const & names); //! Set the parameter names (clears all columns)
 	bool matchesFilter(size_t ind) const;             //!< Returns true if point with given index matches current filter
-	void addFilter(int paramIndex, double value);     //!< Adds a filter on the given column (index), it needs to match the given value; multiple filters are linked via OR
-	void removeFilter(int paramIndex, double value);  //!< Removes the filter on the given column and value.
+	void addFilter(size_t paramIndex, double value);  //!< Adds a filter on the given column (index), it needs to match the given value; multiple filters are linked via OR
+	void removeFilter(size_t paramIndex, double value);//!< Removes the filter on the given column and value.
 	void clearFilter();                               //!< Clear all filters
 	bool filterDefined() const;                       //!< Returns true if a filter is defined on the data
 	double const* paramRange(size_t paramIndex) const;//!< Get the range of the parameter with given index
@@ -69,5 +69,5 @@ protected:
 	std::vector<char> m_inverted;                     //!< whether to invert a feature
 private:
 	void updateRangeInternal(size_t paramIndex);      //!< Update internal range data for parameter paramIndex
-	std::vector<std::pair<int,  double> > m_filters;  //!< collection of filters: each column index/value pair is linked via OR
+	std::vector<std::pair<size_t, double> > m_filters;//!< collection of filters: each column index/value pair is linked via OR
 };
